Change code, add code
Change the way the program determines if a line has a plain URL and has
any trailing spaces, also changing the way inequality signs are added.
This applies to a file with multiple lines.
Add if-construct to create a string made up of space characters with a
length equal to the number of space characters in a line. This is used
to determine if a line has a plain URL, checking if all the line's space
characters appear at the end of a line.
Do not accept a line as a plain URL if it contains inequality signs.
Use 'substr' instead of 'erase' to remove any trailing spaces from a
line.
Use 'std::count' to determine the number of space characters in a line.
Add 'unaltered_file_contents' vector to contain unaltered file contents.
Add 'plain_url_found' variable to indicate if a plain URL is found.
Collect variable declarations together.
Add URL in comment.